Replace GenesisChangeReader with GenesisChange function
